List of improvements: - Fix script to run with sbt - Updated documentation - Remove UseCase to make things simpler
