+++ docker/docker-compose.yml
@@ -6,3 +6,5 @@ services:
     restart: always
     ports:
       - 8080:8080
+    env_file:
+      - ../.env
